Major Build 14 March 2017
Use this version to analyze chronic data. CE:
-Bar plotting and EMG plotting occurs with a subset of channels if
specified
-Fixed numerous cosmetic changes to the bar plots, routing through a new
function which makes a better bar plot with error bars
-Changed the organization of the aggregated data across all sessions for
a day - instead of using a 3D structure with repeats, data is now
organized as 2D structure with no repeats across rows
-PAS_validate_EMG_responses.m was duplicated and the copy was modified
to PAS_validate_EMG_responses.m, for which (1) auto trial validation now
is done for 1SD cutoff instead of 2SD; (2) only the trials which are
common to the automatic screening and the manual validation are taken;
(3) ? Across the channel, only the results of 2 are taken.
-other minor changes.